Fix example regression (#385)
